user router split into controller file
